+++ 1-bootstrap/main.tf
@@ -56,8 +56,8 @@ module "tfstate_bucket" {
 module "tf_cloudbuild_workspace" {
   source   = "terraform-google-modules/bootstrap/google//modules/tf_cloudbuild_workspace"
   version  = "~> 7.0"
+  
   for_each = local.cb_config
-
 
   project_id = var.project_id
   location   = var.location
